requirement.txt
Django==4.2.6
django-filter==23.3
django-silk==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
gunicorn==21.2.0